refactor(cli): unify concurrent pipe drain and route eval through shared runner
Duplicate bounded child-process logic removed from eval/repl.rs.
The shared run_binary_with_timeout in process.rs now uses concurrent
stdout/stderr draining via background threads (ConcurrentChildOutput)
instead of sequential post-exit reads, fixing a latent pipe-buffer
deadlock. hew eval is routed through this shared runner.

Files changed:
- hew-cli/src/process.rs: add ConcurrentChildOutput + ChildPipeReader;
  update run_binary_with_timeout to drain concurrently; remove
  collect_child_output and read_pipe.
- hew-cli/src/eval/repl.rs: remove ~170 lines of duplicate process
  code (run_eval_binary_with_timeout, EvalChildOutput, EvalPipeReader,
  spawn_eval_child, kill_timed_out_eval_child, etc.); route
  run_inprocess_compiled through crate::process::run_binary_with_timeout.

Invariants preserved: ProcessGroup SIGKILL semantics, timeout/kill/reap
flow, BinaryRunOutcome API, all existing callers unaffected.
